L5 again

+++ Services/CharacterServices/CharacterService.cs
@@ -1,9 +1,11 @@
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Data;
 using Dtos.Character;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Models;
 
@@ -13,26 +15,36 @@ namespace Services.CharacterServices
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _context = context;
             _mapper = mapper;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto character)
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character characterNew = _mapper.Map<Character>(character);
+            characterNew.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             await _context.Characters.AddAsync(characterNew);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+            serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacter(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacter()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            List<Character> dbCharacters = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Include(c => c.Weapon)
+            .Include(c => c.CharacterSkills).ThenInclude(cs => cs.Skill)
+            .ToListAsync();
             serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
@@ -40,7 +52,10 @@ namespace Services.CharacterServices
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            Character dbCharacter = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.CharacterSkills).ThenInclude(cs => cs.Skill)
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -51,18 +66,25 @@ namespace Services.CharacterServices
 
             try
             {
-                Character characterNew = await _context.Characters.FirstOrDefaultAsync(c => c.Id == character.Id);
-                characterNew.Name = character.Name;
-                characterNew.HitPoints = character.HitPoints;
-                characterNew.Strength = character.Strength;
-                characterNew.Defence = character.Defence;
-                characterNew.Intelligence = character.Intelligence;
-                characterNew.rpgEnum = character.rpgEnum;
-
-                _context.Characters.Update(characterNew);
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = _mapper.Map<GetCharacterDto>(characterNew);
+                Character characterNew = await _context.Characters
+                    .Include(c => c.User)
+                    .FirstOrDefaultAsync(c => c.Id == character.Id);
+                if (characterNew.User.Id == GetUserId())
+                {
+                    characterNew.Name = character.Name;
+                    characterNew.HitPoints = character.HitPoints;
+                    characterNew.Strength = character.Strength;
+                    characterNew.Defence = character.Defence;
+                    characterNew.Intelligence = character.Intelligence;
+                    characterNew.rpgEnum = character.rpgEnum;
+                    _context.Characters.Update(characterNew);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(characterNew);
+                } else {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
+                
             }
             catch (System.Exception exp)
             {
@@ -78,10 +100,19 @@ namespace Services.CharacterServices
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character characterNew = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(characterNew);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                Character characterNew = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (characterNew != null)
+                {
+                    _context.Characters.Remove(characterNew);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = (_context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                } else {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
+                }
             }
             catch (System.Exception exp)
             {
